chore: remove obsolete code
