--- pkg/cmd/report/templates/code_group.html
@@ -1,55 +1,0 @@
-{{- $icons := .Icons -}}
-<details>
-  <summary>
-    {{ .CodeData.Icon }} {{ .CodeData.Label }}
-    <table align="right">
-      <tr>
-        <td>{{- index .Icons "high" }} {{ len .CumulatedSeverities.High -}}</td>
-        <td>
-          {{- index .Icons "medium" }} {{ len .CumulatedSeverities.Medium -}}
-        </td>
-        <td>{{- index .Icons "low" }} {{ len .CumulatedSeverities.Low -}}</td>
-      </tr>
-    </table>
-    <p>
-      {{- .CumulatedSeverities.TotalAmount }} issue{{- .CumulatedSeverities.TotalAmount | pluralize }} found
-    </p>
-  </summary>
-
-{{ range $severity, $packages := .CumulatedSeverities.Severities }}
-{{ if gt (len $packages) 0 }}
-<details>
-<summary>
-{{- (printf "%s" $severity) | icon }} <b>{{ $severity | severityLabel }}</b> ∙ {{ len $packages }} package{{- len $packages | pluralize }}
-</summary>
-<br />
-
-{{ range $nameVersion, $code := $packages }}
-<details>
-<summary>~ {{ index $icons "package" }} <i>{{ $nameVersion -}}</i></summary>
-
-{{ range $c, $grouped := $code }}
-
-{{- $groupedAmount := $grouped | transitiveAndNonAmount -}}
-`{{- $grouped | extractMessage -}}` ∙ {{ $groupedAmount }} occurrence{{- $groupedAmount | pluralize }}
-
-| Name | Version | Transitive Dependency | More |
-|---|---|---|---|
-{{ range $grouped.Transitive -}}
-{{- $name := index .Metadata "npm_package_name" -}}
-{{- $version := index .Metadata "npm_package_version" -}}
-| {{ $name }} | {{ $version }} | ✔️ | [🔗](https://verdicts.listen.dev/npm/{{- $name -}}/{{- $version -}}) |
-{{ end -}}
-{{- range $grouped.NonTransitive -}}
-| {{ .Pkg }} | {{ .Version }} | | [🔗](https://verdicts.listen.dev/npm/{{- $nameVersion -}}) |
-{{ end }}
-
-{{ end }}
-</details>
-{{ end }}
-</details>
-{{ end }}
-{{ end }}
-</details> 
-
-<hr />
