--- pkg/cmd/report/templates/testdata/actual/README.md
@@ -1,1 +1,0 @@
-TODO
